# apifox/fnmedia-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "飞牛影视（fnOS TrimMedia）服务端 API",
    "version": "1.0.0",
    "description": "FnTube「飞牛 client」实际调用的上游飞牛影视（fnOS TrimMedia）服务端接口集合。\n\n## 服务地址\n默认 baseUrl 形如 `http://IP:端口/v`（含 `/v` 后缀）。用户配置的「服务端地址」若未以 `/v` 结尾，客户端会自动补 `/v` 后重试。\n\n## 鉴权\n所有 `/api/*` 接口需要携带以下请求头：\n- `Authorization`: 登录接口返回的 `data.token`（登录前不携带）。\n- `authx`: 请求签名，格式 `nonce={6位随机数}&timestamp={毫秒时间戳}&sign={md5签名}`。\n\n签名算法（Go 实现见 `client.getAuthx`）：\n```\napiPathStr = \"/v\" + apiPath（例如 /v/api/v1/item/list）\ndataHash   = md5(body)          // GET 且无 body 时为 query string 的编码值\nsign       = md5(secret + \"_\" + apiPathStr + \"_\" + nonce + \"_\" + ts + \"_\" + dataHash + \"_\" + apiKey)\n```\n其中 `apiKey`、`secret` 为客户端内置常量（`16CCEB3D-AB42-077D-36A1-F355324E4237` / `NDzZTVxnRKP8Z0jXg1VAMonaG8akvh`）。\n\n## 统一响应结构\n所有 `/api/*` 接口返回统一包裹：`{ \"code\": int, \"msg\": string, \"data\": ... }`，`code == 0` 表示成功。\n"
  },
  "servers": [
    {
      "url": "{{baseUrl}}",
      "description": "飞牛影视服务端地址（含 /v），例如 http://127.0.0.1:5666/v"
    },
    {
      "url": "{{hostRoot}}",
      "description": "主机根地址（不含 /v），仅访问码校验接口使用，例如 http://127.0.0.1:5666"
    }
  ],
  "security": [
    {
      "Authorization": [],
      "authx": []
    }
  ],
...
  "paths": {
...
    "/api/v1/item/{guid}": {
      "get": {
        "tags": ["媒体项"],
        "summary": "媒体详情",
        "operationId": "itemDetail",
        "description": "获取单个媒体条目详情。",
        "parameters": [
          {
            "name": "guid",
            "in": "path",
            "required": true,
            "description": "媒体条目 GUID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "媒体详情",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 0 },
                    "msg": { "type": "string", "example": "" },
                    "data": { "$ref": "#/components/schemas/Item" }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["媒体项"],
        "summary": "删除媒体",
        "operationId": "deleteItem",
        "description": "删除媒体条目。delete_file=1 删除媒体文件，0 仅从媒体库移除。",
        "parameters": [
          {
            "name": "guid",
            "in": "path",
            "required": true,
            "description": "媒体条目 GUID",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeleteItemRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "删除结果",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EmptyEnvelope" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/season/list/{tvGuid}": {
      "get": {
        "tags": ["剧集"],
        "summary": "季列表",
        "operationId": "seasonList",
        "description": "获取指定剧集的季列表。data 为 Item 数组（每项为 Season 类型）。",
        "parameters": [
          {
            "name": "tvGuid",
            "in": "path",
            "required": true,
            "description": "剧（TV）GUID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "季列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 0 },
                    "msg": { "type": "string", "example": "" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Item" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/episode/list/{seasonGuid}": {
      "get": {
        "tags": ["剧集"],
        "summary": "剧集列表",
        "operationId": "episodeList",
        "description": "获取指定季的剧集列表。data 为 Item 数组（每项为 Episode 类型）。",
        "parameters": [
          {
            "name": "seasonGuid",
            "in": "path",
            "required": true,
            "description": "季 GUID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "剧集列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "integer", "example": 0 },
                    "msg": { "type": "string", "example": "" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Item" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Authorization": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "登录接口返回的 data.token（登录/访问码校验接口可不携带）"
      },
      "authx": {
        "type": "apiKey",
        "in": "header",
        "name": "authx",
        "description": "请求签名，格式 nonce=xxx&timestamp=xxx&sign=xxx，算法见 info.description"
      }
    },
    "schemas": {
...
      "EmptyEnvelope": {
        "type": "object",
        "description": "统一响应包裹（无具名 data 场景）",
        "properties": {
          "code": { "type": "integer", "description": "0 表示成功", "example": 0 },
          "msg": { "type": "string", "example": "" },
          "data": { "description": "业务数据，随接口而定" }
        }
      },
...
      "DeleteItemRequest": {
        "type": "object",
        "properties": {
          "delete_file": { "type": "integer", "description": "1 删除媒体文件，0 仅从媒体库移除", "example": 0 },
          "media_guids": { "type": "array", "items": { "type": "string" }, "description": "关联媒体 GUID 列表" }
        }
      },
...
      "MediaStream": {
        "type": "object",
        "properties": {
          "resolutions": { "type": "array", "items": { "type": "string" } },
          "audio_type": { "type": "array", "items": { "type": "string" } },
          "color_range_type": { "type": "array", "items": { "type": "string" } }
        }
      },
      "Item": {
        "type": "object",
        "description": "媒体条目（对齐 /item 接口响应字段）",
        "properties": {
          "guid": { "type": "string" },
          "ancestor_guid": { "type": "string" },
          "type": { "type": "string", "description": "Movie / TV / Season / Episode / Video / Directory" },
          "tv_title": { "type": "string" },
          "parent_title": { "type": "string" },
          "title": { "type": "string" },
          "original_title": { "type": "string" },
          "overview": { "type": "string" },
          "poster": { "type": "string" },
          "backdrops": { "type": "string" },
          "posters": { "type": "string" },
          "logos": { "type": "string" },
          "douban_id": { "type": "integer" },
          "imdb_id": { "type": "string" },
          "trim_id": { "type": "string", "description": "飞牛 TMDB ID（tt* 表示电影，tm* 表示剧集）" },
          "release_date": { "type": "string" },
          "air_date": { "type": "string" },
          "vote_average": { "type": "string" },
          "season_number": { "type": "integer" },
          "episode_number": { "type": "integer" },
          "duration": { "type": "integer", "description": "片长（秒）" },
          "ts": { "type": "integer", "description": "已播放（秒）" },
          "watched": { "type": "integer", "description": "1 表示已看完" },
          "poster_width": { "type": "integer" },
          "poster_height": { "type": "integer" },
          "genres": { "type": "array", "items": { "type": "integer" } },
          "runtime": { "type": "integer", "description": "片长（分钟）" },
          "production_countries": { "type": "array", "items": { "type": "string" } },
          "is_favorite": { "type": "integer" },
          "is_watched": { "type": "integer" },
          "watched_ts": { "type": "integer" },
          "number_of_episodes": { "type": "integer" },
          "local_number_of_episodes": { "type": "integer" },
          "local_number_of_seasons": { "type": "integer" },
          "can_play": { "type": "integer" },
          "play_error": { "type": "string" },
          "parent_guid": { "type": "string" },
          "ancestor_name": { "type": "string" },
          "ancestor_category": { "type": "string" },
          "play_item_guid": { "type": "string" },
          "logic_type": { "type": "integer" },
          "media_stream": { "$ref": "#/components/schemas/MediaStream" }
        }
      },
...
    }
  }
}